# wms_api/api-contracts.yaml
openapi: 3.0.3
info:
  title: WMS API - Warehouse Management System
  description: |
    API контракты для системы управления складом (WMS).
    Включает управление товарами, заказами, размещением, пользователями и логистикой доставки.
  version: 1.0.0
  contact:
    name: WMS Development Team
servers:
  - url: http://localhost:3002/api
    description: Development server (WMS API)
  - url: http://localhost:3001/api
    description: Mock server (wms_site/server)
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Выход из системы
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление токена
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Токен обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  # ---------------------------------------------------------------------------
  # USERS
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
...
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserInfo'
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        role:
          type: string
        email:
          type: string

    # --- Users ---
...